# docs/swagger.yaml
basePath: /
definitions:
...
  models.Message:
    properties:
      chat_id:
        type: integer
      content:
        type: string
      id:
        type: integer
      sended:
        type: string
      user_id:
        type: integer
    type: object
  models.Response:
    properties:
      content: {}
      message:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: pidorpidorasi
  title: chat
  version: "1.1"
paths:
...
  /message:
    put:
      consumes:
      - application/json
      description: Update a message with new content
      operationId: update-message
      parameters:
      - description: Message object
        in: body
        name: message
        required: true
        schema:
          $ref: '#/definitions/models.Message'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Update a message
      tags:
      - messages
  /message/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a message with a specified ID
      operationId: delete-message
      parameters:
      - description: Message ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Delete a message
      tags:
      - messages
  /messages/{chat}/{limit}:
    get:
      consumes:
      - application/json
      description: Retrieve messages for a specific chat with a specified limit
      operationId: get-chat-messages
      parameters:
      - description: Chat ID
        in: path
        name: chat
        required: true
        type: integer
      - description: Limit of messages to retrieve
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get chat messages
      tags:
      - messages
  /messages/{limit}:
    get:
      consumes:
      - application/json
      description: Retrieve messages with a specified limit
      operationId: get-messages
      parameters:
      - description: Limit of messages to retrieve
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get messages
      tags:
      - messages
  /messages/{user}/{limit}:
    get:
      consumes:
      - application/json
      description: Retrieve messages for a specific user with a specified limit
      operationId: get-user-messages
      parameters:
      - description: User ID
        in: path
        name: user
        required: true
        type: integer
      - description: Limit of messages to retrieve
        in: path
        name: limit
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get user messages
      tags:
      - messages
  /upload/photo:
    post:
      consumes:
      - multipart/form-data
      description: Uploads a photo with the specified timing, user ID, chatroom ID,
        and photo file
      operationId: uploadPhoto
      parameters:
      - description: Timing
        in: formData
        name: timing
        required: true
        type: string
      - description: User ID
        in: formData
        name: user_id
        required: true
        type: integer
      - description: Chatroom ID
        in: formData
        name: chatroom_id
        required: true
        type: integer
      - description: Photo file
        in: formData
        name: photo
        required: true
        type: file
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
        "500":
          description: Internal Server Error
          schema:
            $ref: '#/definitions/models.Response'
      summary: Uploads a photo
      tags:
      - messages
...
  /user/{id}:
    get:
      consumes:
      - application/json
      description: Retrieve a user with a specified ID
      operationId: get-user
      parameters:
      - description: User ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/models.Response'
        "400":
          description: Bad Request
          schema:
            $ref: '#/definitions/models.Response'
      summary: Get a user
      tags:
      - user
...
securityDefinitions:
  JWTAuth:
    type: basic
swagger: "2.0"